producer/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
python-dotenv==1.0.0
backoff==2.2.1
schedule==1.2.0